refactor: add testimonial card component
